python_predict_client/requirements.txt
grpcio>=0.15.0
grpcio-tools>=0.15.0
tensorflow-serving-api